+++ src/models/testimonial.js
@@ -1,25 +1,26 @@
 'use strict';
 const {
-  Model
+    Model
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-  class Testimonial extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
+    class Testimonial extends Model {
+        /**
+         * Helper method for defining associations.
+         * This method is not a part of Sequelize lifecycle.
+         * The `models/index` file will call this method automatically.
+         */
+        static associate(models) {
+            Testimonial.belongsTo(models.User, { as: 'user' })
+        }
     }
-  }
-  Testimonial.init({
-    name: DataTypes.STRING,
-    image: DataTypes.STRING,
-    content: DataTypes.STRING
-  }, {
-    sequelize,
-    tableName: 'testimonials',
-  });
-  return Testimonial;
+    Testimonial.init({
+        name: DataTypes.STRING,
+        image: DataTypes.STRING,
+        content: DataTypes.STRING,
+        // userId: DataTypes.INTEGER
+    }, {
+        sequelize,
+        tableName: 'testimonials',
+    });
+    return Testimonial;
 };
