+++ src/models/user.js
@@ -1,29 +1,30 @@
 'use strict';
 const {
-  Model
+    Model
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-  class User extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      User.belongsTo(models.Role, {as: 'role'});
-    }
-  };
-  User.init({
-    firstName: DataTypes.STRING,
-    lastName: DataTypes.STRING,
-    email: DataTypes.STRING,
-    image: DataTypes.STRING,
-    password: DataTypes.STRING,
-    roleId: DataTypes.INTEGER,
-    deletedAt: DataTypes.DATE
-  }, {
-    sequelize,
-    tableName: 'users'
-  });
-  return User;
+    class User extends Model {
+        /**
+         * Helper method for defining associations.
+         * This method is not a part of Sequelize lifecycle.
+         * The `models/index` file will call this method automatically.
+         */
+        static associate(models) {
+            User.belongsTo(models.Role, { as: 'role' });
+            User.hasOne(models.Testimonial, { as: 'testimonial' })
+        }
+    };
+    User.init({
+        firstName: DataTypes.STRING,
+        lastName: DataTypes.STRING,
+        email: DataTypes.STRING,
+        image: DataTypes.STRING,
+        password: DataTypes.STRING,
+        roleId: DataTypes.INTEGER,
+        deletedAt: DataTypes.DATE
+    }, {
+        sequelize,
+        tableName: 'users'
+    });
+    return User;
 };
